fix(import): Use addon-test-support and fix PactEnabled mirage in dev
