+++ addon-test-support/ember-cli-pact/-private/testing/index.js
@@ -3,7 +3,8 @@
 import require from 'require';
 import { assert } from '@ember/debug';
 import { getOwner } from '@ember/application';
-import { getContext } from '@ember/test-helpers';
+// import { getContext } from '@ember/test-helpers';
+import { getContext } from 'ember-test-helpers';
 import { uploadInteraction, finalize } from './upload';
 
 export function setupPact(hooks = {}, options = {}) {
@@ -74,9 +75,15 @@ function setupProvider(context, options, testName) {
 }
 
 function loadMockProvider(context, options) {
-  let { modulePrefix } = getConfig(context);
   let name = getConfigValue(context, options, 'mockProvider');
-  return require(`${modulePrefix}/tests/helpers/pact-providers/${name}`).default;
+  try {
+    // Look in the default provider location (from addon-test-support)
+    return require(`ember-cli-pact/pact-providers/${name}`).default;
+  } catch (error) {
+    // Fall back to looking for a custom provider in the helpers directory
+    let { modulePrefix } = getConfig(context);
+    return require(`${modulePrefix}/tests/helpers/pact-providers/${name}`).default;
+  }
 }
 
 function assertSingleConsumerName(context, options) {
