+++ addon-test-support/ember-cli-pact/index.js
@@ -1,4 +1,5 @@
-import { getContext } from '@ember/test-helpers';
+// import { getContext } from '@ember/test-helpers'; TODO: Newer test-helpers
+import { getContext } from 'ember-test-helpers';
 
 export { setupPact } from './-private/testing';
 export { registerProviderState as providerState } from './-private/provider-states';
