+++ addon-test-support/ember-cli-pact/mock-provider/index.js
@@ -1,6 +1,7 @@
 import { assert } from '@ember/debug';
 import { run } from '@ember/runloop';
-import { getContext } from '@ember/test-helpers';
+// import { getContext } from '@ember/test-helpers';
+import { getContext } from 'ember-test-helpers';
 import { Promise } from 'rsvp';
 
 import Interaction from 'ember-cli-pact/-private/interaction';
