+++ tests/helpers/pact-providers/custom.js
@@ -1,0 +1,2 @@
+// Ok, not so custom, but you get the idea ;)
+export {default} from 'ember-cli-pact/mock-provider/mirage'
